+++ packages/reactabular-dnd/CHANGELOG.md
@@ -1,7 +1,7 @@
 7.0.0 / 2016-xx-xx
 ==================
 
-  * Breaking - Rework `moveRows` interface. Instead of `moveRows(rows, { sourceRowId, targetRowId }) => rows` it's `moveRows({ rows, sourceRowId, targetRowId }) => { rows, sourceIndex, targetIndex }` now.
+  * Breaking - Rework `moveRows` interface. Instead of `moveRows(rows, { sourceRowId, targetRowId }) => rows` it's `moveRows({ rows, sourceRowId, targetRowId }) => rows` now.
 
 2.0.0 / 2016-08-16
 ==================